requirements.txt
Packages:
pandas_datareader
alpaca_trade_api
pyEX

Accounts
iex account
alpaca account
